Change size type of ilot int to float

+++ src/Entity/Ilots.php
@@ -30,7 +30,7 @@ class Ilots
     private $name;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $size;
 
@@ -91,12 +91,12 @@ class Ilots
         return $this;
     }
 
-    public function getSize(): ?int
+    public function getSize(): ?float
     {
         return $this->size;
     }
 
-    public function setSize(int $size): self
+    public function setSize(float $size): self
     {
         $this->size = $size;
